ml/requirements.txt
keras>=2.0.4,<=2.1.0
tensorflow
Pillow
coremltools
pandas
h5py
